requirements.txt
Django~=4.0.1
django-debug-toolbar~=3.2.4
environs~=9.5.0
django-ckeditor~=6.2.0
django-simple-captcha~=0.5.14